requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.1
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
gunicorn==21.2.0
requests==2.31.0 